requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-environ==0.9.0
idna==3.4
psycopg2==2.9.5
pycodestyle==2.10.0
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.13
